Transfered from pactice folder
Font selection on message is working … might want to go link out to
google instead.
Building the image usrls in pick a shape, pick a metal, and pic a size.
All clicks are working and returning relevant parts of the url. Now to
put them together as a whole url to get placed on the preview
background.