Fixed the print button error.
I just want to point out that you will see 3 errors in the css
validation. Those are hacks for IE 6/7 and can not be avoided if I want
my pages to display properly in IE 6 and 7. Some day we won't care so
much but I know a lot of large corporations still test in IE 6 and 7 …
banks and such … Thank you 3 down 1 to go